legacy/python/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "fs-explorer"
version = "0.1.0"
description = "Explore and understand your filesystem better with AI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docling>=2.55.0",
    "fastapi>=0.115.0",
    "google-genai>=1.55.0",
    "langextract>=1.0.0",
    "llama-index-workflows>=2.11.5",
    "openai>=2.32.0",
    "pgvector>=0.3.6",
    "pymupdf>=1.24.0",
    "pymupdf4llm",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.7",
    "rich>=13.0.0",
    "typer>=0.12.5,<0.20.0",
    "uvicorn>=0.34.0",
    "websockets>=14.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.9",
    "ty>=0.0.1a33",
]

[project.scripts]
explore = "fs_explorer.main:app"
explore-ui = "fs_explorer.server:run_server"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
